DataSet refactoring: capabilities, alignment, mmap, empad reader, merge raw readers, ...
Fixes #207; refs #252, #188

+++ tests/utils.py
@@ -41,7 +41,7 @@ class MemoryDataSet(DataSet):
         self.sig_dims = sig_dims
         self._meta = DataSetMeta(
             shape=self.shape,
-            dtype=self.data.dtype,
+            raw_dtype=self.data.dtype,
         )
 
     @property
